feat(examples): Improve http-c
Improve documentation of native `http-c`. Move it from `examples/`
directory to `native/` directory.
