--- examples/http-c/Makefile.uk
@@ -1,3 +1,0 @@
-$(eval $(call addlib,apphttpreply))
-
-APPHTTPREPLY_SRCS-y += $(APPHTTPREPLY_BASE)/main.c
